Shifted to use a generic Tag<Ty, Over> Type in place of a Beacon<T> to support new type checker
