+++ flywheel-compiler/src/utils/mod.rs
@@ -5,6 +5,8 @@ pub mod once_map;
 pub mod sync_interner;
 pub mod located;
 pub mod pretty_tree;
+pub mod tag;
+pub mod beacon;
 
 pub use keyed_map::*;
 pub use once_map::*;
